// Apps/Client/src/domain/interfaces/identity/me.ts
export interface IMeResponse {
   first_name: string;
   last_name: string;
   phone_number: string;
   birth_date: string;
   image_id: string;
   image_url: string;
   default_address_label: string;
   default_contact_name: string;
   default_contact_phone_number: string;
   default_address_line: string;
   default_address_district: string;
   default_address_province: string;
   default_address_country: string;
}